Stop Click Jacking
Click Jacking is when someone puts your site in an iFrame and then puts hidden elements over your elements. An example would be if someone bought girlscodelincoln.org and then added a element over the submit button that would look to see if they are logged into their bank account and then it could transfer money that way.